% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratr.R
\name{rat_house}
\alias{rat_house}
\title{Get a rat parser}
\usage{
rat_house(path = "rat")
}
\arguments{
\item{path}{the path to the rat compiler}
}
\value{
A function that parses rat code
}
\description{
Get a rat parser
}

